social_media_core/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-currentuser==0.5.3
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
